requirements.txt
asgiref==3.8.1
celery==5.3.6
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.5
cron-descriptor==1.4.5
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.1.4
django-allauth==0.13.0
django-celery-beat==2.8.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-storages==1.14.2
django-timezone-field==7.1
django-widget-tweaks==1.5.0
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.4
Jinja2==3.1.2
kombu==5.5.3
\
Pillow==10.0.1
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
redis==6.0.0
requests==2.31.0

tzdata==2025.2

